Payment with Card shifted
