+++ src/features/public/Donations/components/PaymentDetails.tsx
@@ -12,13 +12,12 @@ import React, { ReactElement } from 'react';
 import Sugar from 'sugar';
 import CreditCard from '../../../../assets/images/icons/donation/CreditCard';
 import BackArrow from '../../../../assets/images/icons/headerIcons/BackArrow';
-import { getCountryDataBy } from '../../../../utils/countryUtils';
 import { getCardBrand } from '../../../../utils/stripeHelpers';
 import PaymentProgress from '../../../common/ContentLoaders/Donations/PaymentProgress';
 import AnimatedButton from '../../../common/InputTypes/AnimatedButton';
 import { PaymentDetailsProps } from './../../../common/types/donations';
 import styles from './../styles/PaymentDetails.module.scss';
-import { createDonation, payDonation } from './treeDonation/PaymentFunctions';
+import { payWithCard } from './treeDonation/PaymentFunctions';
 
 const FormControlNew = withStyles({
   root: {
@@ -132,12 +131,10 @@ function PaymentDetails({
     }
 
     let paymentMethod: any;
-    let error;
 
     if (paymentType === 'CARD') {
       const cardElement = elements!.getElement(CardNumberElement);
       cardElement!.on('change', ({ error }) => {
-        const displayError = document.getElementById('payment-errors');
         if (error) {
           setPaymentError(error.message);
           return;
@@ -171,132 +168,23 @@ function PaymentDetails({
       paymentMethod = payload.paymentMethod;
       // Add payload error if failed
     }
-    setIsPaymentProcessing(true);
-    let countryCode = getCountryDataBy(
-      'countryName',
-      contactDetails.country.toString()
-    )
-      ? getCountryDataBy('countryName', contactDetails.country.toString())
-          .countryCode
-      : null;
-    let createDonationData = {
-      type: 'trees',
-      project: project.id,
-      treeCount: treeCount,
-      amount: treeCost * treeCount,
-      currency: currency,
-      donor: {
-        firstname: contactDetails.firstName,
-        lastname: contactDetails.lastName,
-        email: contactDetails.email,
-        address: contactDetails.address,
-        zipCode: contactDetails.zipCode,
-        city: contactDetails.city,
-        country: countryCode,
-      },
+    const payWithCardProps = {
+      setDonationStep,
+      setIsPaymentProcessing,
+      contactDetails,
+      project,
+      currency,
+      treeCost,
+      treeCount,
+      giftDetails,
+      isGift,
+      setPaymentError,
+      stripe,
+      paymentSetup,
+      window,
+      paymentMethod,
     };
-    let gift = {
-      gift: {
-        type: 'invitation',
-        recipientName: giftDetails.recipientName,
-        recipientEmail: giftDetails.email,
-        message: giftDetails.giftMessage,
-      },
-    };
-    if (isGift) {
-      createDonationData = {
-        ...createDonationData,
-        ...gift,
-      };
-    }
-    createDonation(createDonationData)
-      .then((res) => {
-        // Code for Payment API
-
-        if (res.code === 400) {
-          setIsPaymentProcessing(false);
-          setPaymentError(res.message);
-          return;
-        } else {
-          const payDonationData = {
-            paymentProviderRequest: {
-              account: paymentSetup.gateways.stripe.account,
-              gateway: 'stripe_pi',
-              source: {
-                id: paymentMethod.id,
-                object: 'payment_method',
-              },
-            },
-          };
-
-          payDonation(payDonationData, res.id)
-            .then(async (res) => {
-              if (res.status === 'failed') {
-                setIsPaymentProcessing(false);
-                setPaymentError(res.message);
-                return;
-              } else {
-                if (res.paymentStatus === 'success') {
-                  setIsPaymentProcessing(false);
-                  setDonationStep(4);
-                } else if (res.status === 'action_required') {
-                  const clientSecret =
-                    res.response.payment_intent_client_secret;
-                  const donationID = res.id;
-                  const stripe = window.Stripe(
-                    process.env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY!,
-                    {
-                      stripeAccount: res.response.account,
-                    }
-                  );
-                  if (stripe) {
-                    await stripe.handleCardAction(clientSecret).then((res) => {
-                      if (res.error) {
-                        setIsPaymentProcessing(false);
-                        setPaymentError(res.error.message);
-                      } else {
-                        const payDonationData = {
-                          paymentProviderRequest: {
-                            account: paymentSetup.gateways.stripe.account,
-                            gateway: 'stripe_pi',
-                            source: {
-                              id: res.paymentIntent.id,
-                              object: 'payment_intent',
-                            },
-                          },
-                        };
-                        payDonation(payDonationData, donationID).then((res) => {
-                          if (res.paymentStatus === 'success') {
-                            setIsPaymentProcessing(false);
-                            setDonationStep(4);
-                          } else {
-                            setIsPaymentProcessing(false);
-                            setPaymentError(res.error.message);
-                          }
-                        });
-                      }
-                    });
-                  }
-                }
-              }
-            })
-            .catch((error) => {
-              setIsPaymentProcessing(false);
-              setPaymentError(error.message);
-              return;
-            }); // Add Catch if pay donation failes
-        }
-      })
-      .catch((error) => {
-        setIsPaymentProcessing(false);
-        setPaymentError(error.message);
-        return;
-      }); // Add Catch if create donation failes
-    if (error) {
-      setIsPaymentProcessing(false);
-      setPaymentError(error.message);
-      return;
-    }
+    payWithCard({ ...payWithCardProps });
   };
 
   return isPaymentProcessing ? (
